adding function for seperating answers from aops page source code
first function takes a string (in this case one of the answer choices) as it will be separated by "textbf" in any of the formats that appear on the aops source code and turn it into desired format (letter space content). questions with fractions are excluded. second function simply repeats this for every string in a list and adds it to a new list.